Add Windows to Evergreen
Changed all scripts to use /bin/bash instead of relying on /bin/sh

+++ src/mongocrypt-key-broker.c
@@ -285,7 +285,7 @@ _mongocrypt_key_broker_filter (_mongocrypt_key_broker_t *kb,
 
       key_str = bson_strdup_printf ("%d", i++);
       _mongocrypt_buffer_append (
-         &iter->key_id, &_id_in, key_str, strlen (key_str));
+         &iter->key_id, &_id_in, key_str, (uint32_t) strlen (key_str));
 
       bson_free (key_str);
    }
